chore: refactor lsp
